Fix Git#version with groups
When using groups cfg.repo returns a config struct, not a string that can be fed to File.dirname
It seems the point of the code is to just retrieve the correct repo for the group. So it does that now.

Fixes a "node not found" on the versions page.